[BY-211] Refactor: make nickname file
